Move filter from bookForm to BookList
